File update
git-svn-id: https://plugins.svn.wordpress.org/wptouch/tags/3.7.7@1147597 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ core/class-wptouch-pro.php
@@ -833,7 +833,7 @@ class WPtouchProThree {
 					}
 
 					foreach( $url_list as $url ) {
-						if ( $url == '/' || strpos( trim( $server_url, '\/' ), trim( $url, '\/' ) ) !== false ) {
+						if ( ( $url == '/' && $server_url == '/' ) || strpos( trim( $server_url, '\/' ), trim( $url, '\/' ) ) !== false ) {
 							if ( $settings->url_filter_behaviour == 'exclude_urls' ) { // Excluding URLs - kill mobile if the URL is matched
 								$block_mobile = true;
 							} elseif( $settings->url_filter_behaviour == 'exclusive_urls' ) { // Exclusive URLs - kill mobile if the URL is *not* matched
